# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocalSpot Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout/create-payment-intent:
    post:
      operationId: createPaymentIntent
      tags: [checkout]
      summary: Create a Stripe PaymentIntent for a spot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentBody"
      responses:
        "200":
          description: PaymentIntent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/confirm:
    post:
      operationId: confirmPayment
      tags: [checkout]
      summary: Confirm payment and mark spot as paid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmPaymentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreatePaymentIntentBody:
      type: object
      properties:
        spotId:
          type: integer
      required:
        - spotId
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        spotId:
          type: integer
        amount:
          type: integer
        businessName:
          type: ["string", "null"]
        size:
          type: string
          enum: [xl, large, medium, small]
      required:
        - clientSecret
        - spotId
        - amount
        - size
    ConfirmPaymentBody:
      type: object
      properties:
        paymentIntentId:
          type: string
        spotId:
          type: integer
      required:
        - paymentIntentId
        - spotId
    ConfirmPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        orderId:
          type: integer
      required:
        - success
        - orderId
...